[#5268] Configured Objects API registration for the children component
