+++ src/openforms/registrations/contrib/objects_api/tests/files/vcr_cassettes/ObjectsAPIBackendV2Tests/test_submission_with_objects_api_v2_with_payment_attributes.yaml
@@ -29,11 +29,11 @@ interactions:
       Cross-Origin-Opener-Policy:
       - same-origin
       Date:
-      - Mon, 18 Aug 2025 11:46:49 GMT
+      - Fri, 19 Sep 2025 14:07:46 GMT
       Referrer-Policy:
       - same-origin
       Server:
-      - nginx/1.29.0
+      - nginx/1.27.4
       Vary:
       - origin
       X-Content-Type-Options:
@@ -70,7 +70,7 @@ interactions:
     uri: http://localhost:8002/api/v2/objects
   response:
     body:
-      string: '{"url":"http://objects-web:8000/api/v2/objects/7fde9f3a-e8f5-4598-9421-ef57992dad11","uuid":"7fde9f3a-e8f5-4598-9421-ef57992dad11","type":"http://objecttypes-web:8000/api/v2/objecttypes/8e46e0a5-b1b4-449b-b9e9-fa3cea655f48","record":{"index":1,"typeVersion":3,"data":{"submission_payment_completed":true,"submission_payment_amount":40.0,"submission_payment_public_ids":["foo","bar"],"submission_provider_payment_ids":["123456","654321"]},"geometry":null,"startAt":"2024-03-19","endAt":null,"registrationAt":"2025-08-18","correctionFor":null,"correctedBy":null}}'
+      string: '{"url":"http://objects-web:8000/api/v2/objects/ca5177d1-aed6-4e24-a3dd-ed835a413959","uuid":"ca5177d1-aed6-4e24-a3dd-ed835a413959","type":"http://objecttypes-web:8000/api/v2/objecttypes/8e46e0a5-b1b4-449b-b9e9-fa3cea655f48","record":{"index":1,"typeVersion":3,"data":{"submission_payment_completed":true,"submission_payment_amount":40.0,"submission_payment_public_ids":["foo","bar"],"submission_provider_payment_ids":["123456","654321"]},"geometry":null,"startAt":"2024-03-19","endAt":null,"registrationAt":"2025-09-19","correctionFor":null,"correctedBy":null}}'
     headers:
       Allow:
       - GET, POST, HEAD, OPTIONS
@@ -85,13 +85,13 @@ interactions:
       Cross-Origin-Opener-Policy:
       - same-origin
       Date:
-      - Mon, 18 Aug 2025 11:46:49 GMT
+      - Fri, 19 Sep 2025 14:07:47 GMT
       Location:
-      - http://localhost:8002/api/v2/objects/7fde9f3a-e8f5-4598-9421-ef57992dad11
+      - http://localhost:8002/api/v2/objects/ca5177d1-aed6-4e24-a3dd-ed835a413959
       Referrer-Policy:
       - same-origin
       Server:
-      - nginx/1.29.0
+      - nginx/1.27.4
       Vary:
       - origin
       X-Content-Type-Options:
